+++ lib/getch/filesystem/zfs/encrypt/partition.rb
@@ -79,31 +79,33 @@ module Getch
               # https://openzfs.github.io/openzfs-docs/Getting%20Started/Ubuntu/Ubuntu%2020.04%20Root%20on%20ZFS.html
               @log.info("Creating boot pool on #{@pool_name}")
               exec("zpool create -f \\
-              -o ashift=#{ashift} -d \\
-              -o feature@async_destroy=enabled \\
-              -o feature@bookmarks=enabled \\
-              -o feature@embedded_data=enabled \\
-              -o feature@empty_bpobj=enabled \\
-              -o feature@enabled_txg=enabled \\
-              -o feature@extensible_dataset=enabled \\
-              -o feature@filesystem_limits=enabled \\
-              -o feature@hole_birth=enabled \\
-              -o feature@large_blocks=enabled \\
-              -o feature@lz4_compress=enabled \\
-              -o feature@spacemap_histogram=enabled \\
-              -O acltype=posixacl -O canmount=off -O compression=lz4 \\
-              -O devices=off -O normalization=formD -O atime=off -O xattr=sa \\
-              -O mountpoint=/boot -R #{MOUNTPOINT} \\
-              #{@boot_pool_name} #{@dev_boot}
-                   ")
+                -o ashift=#{ashift} -d \\
+                -o feature@async_destroy=enabled \\
+                -o feature@bookmarks=enabled \\
+                -o feature@embedded_data=enabled \\
+                -o feature@empty_bpobj=enabled \\
+                -o feature@enabled_txg=enabled \\
+                -o feature@extensible_dataset=enabled \\
+                -o feature@filesystem_limits=enabled \\
+                -o feature@hole_birth=enabled \\
+                -o feature@large_blocks=enabled \\
+                -o feature@lz4_compress=enabled \\
+                -o feature@spacemap_histogram=enabled \\
+                -O acltype=posixacl -O canmount=off -O compression=lz4 \\
+                -O devices=off -O normalization=formD -O atime=off -O xattr=sa \\
+                -O mountpoint=/boot -R #{MOUNTPOINT} \\
+                #{@boot_pool_name} #{@dev_boot}
+              ")
             end
 
             exec("zpool create -f -o ashift=#{ashift} \\
-            -O acltype=posixacl -O canmount=off -O compression=lz4 \\
-            -O dnodesize=auto -O normalization=formD -O atime=off \\
-            -O xattr=sa -O mountpoint=/ -R #{MOUNTPOINT} \\
-            #{@pool_name} #{@dev_root}
-                 ")
+              -O encryption=aes-256-gcm \\
+              -O keylocation=prompt -O keyformat=passphrase \\
+              -O acltype=posixacl -O canmount=off -O compression=lz4 \\
+              -O dnodesize=auto -O normalization=formD -O atime=off \\
+              -O xattr=sa -O mountpoint=/ -R #{MOUNTPOINT} \\
+              #{@pool_name} #{@dev_root}
+            ")
 
             add_datasets
           end